Moved ostream operators for carla::geo types
... to allow usage also from carla::log() calls
