+++ LibCarla/source/carla/geom/Velocity.h
@@ -113,5 +113,10 @@ namespace geom {
 #endif // LIBCARLA_INCLUDED_FROM_UE4
   };
 
+  inline std::ostream &operator<<(std::ostream &out, const Velocity &vector3D) {
+    carla::geom::WriteVector3D(out, "Velocity", vector3D);
+    return out;
+  }
+
 } // namespace geom
 } // namespace carla
