+++ PythonAPI/carla/source/libcarla/Geom.cpp
@@ -20,95 +20,6 @@
 #include <boost/python/implicit.hpp>
 #include <boost/python/suite/indexing/vector_indexing_suite.hpp>
 
-#include <ostream>
-
-namespace carla {
-namespace geom {
-
-  template <typename T>
-  static void WriteVector2D(std::ostream &out, const char *name, const T &vector2D) {
-    out << name
-        << "(x=" << std::to_string(vector2D.x)
-        << ", y=" << std::to_string(vector2D.y) << ')';
-  }
-
-  template <typename T>
-  static void WriteVector3D(std::ostream &out, const char *name, const T &vector3D) {
-    out << name
-        << "(x=" << std::to_string(vector3D.x)
-        << ", y=" << std::to_string(vector3D.y)
-        << ", z=" << std::to_string(vector3D.z) << ')';
-  }
-
-  std::ostream &operator<<(std::ostream &out, const Vector2D &vector2D) {
-    WriteVector2D(out, "Vector2D", vector2D);
-    return out;
-  }
-
-  std::ostream &operator<<(std::ostream &out, const Vector3D &vector3D) {
-    WriteVector3D(out, "Vector3D", vector3D);
-    return out;
-  }
-
-  std::ostream &operator<<(std::ostream &out, const Location &location) {
-    WriteVector3D(out, "Location", location);
-    return out;
-  }
-
-  std::ostream &operator<<(std::ostream &out, const Velocity &vector3D) {
-    WriteVector3D(out, "Velocity", vector3D);
-    return out;
-  }
-
-  std::ostream &operator<<(std::ostream &out, const AngularVelocity &vector3D) {
-    WriteVector3D(out, "AngularVelocity", vector3D);
-    return out;
-  }
-
-  std::ostream &operator<<(std::ostream &out, const Acceleration &vector3D) {
-    WriteVector3D(out, "Acceleration", vector3D);
-    return out;
-  }
-
-  std::ostream &operator<<(std::ostream &out, const Rotation &rotator) {
-    out << "Rotation(pitch=" << std::to_string(rotator.pitch)
-        << ", yaw=" << std::to_string(rotator.yaw)
-        << ", roll=" << std::to_string(rotator.roll) << ')';
-    return out;
-  }
-
-  std::ostream &operator<<(std::ostream &out, const Quaternion &quaternion) {
-    out << "Quaternion"
-        << "(x=" << std::to_string(quaternion.x)
-        << ", y=" << std::to_string(quaternion.y)
-        << ", z=" << std::to_string(quaternion.z)
-        << ", w=" << std::to_string(quaternion.w) << ')';
-    return out;
-  }
-
-  std::ostream &operator<<(std::ostream &out, const Transform &transform) {
-    out << "Transform(" << transform.location << ", " << transform.rotation << ')';
-    return out;
-  }
-
-  std::ostream &operator<<(std::ostream &out, const BoundingBox &box) {
-    out << "BoundingBox(" << box.location << ", ";
-    WriteVector3D(out, "Extent", box.extent);
-    out << ", " << box.rotation;
-    out << ')';
-    return out;
-  }
-
-  std::ostream &operator<<(std::ostream &out, const GeoLocation &geo_location) {
-    out << "GeoLocation(latitude=" << std::to_string(geo_location.latitude)
-        << ", longitude=" << std::to_string(geo_location.longitude)
-        << ", altitude=" << std::to_string(geo_location.altitude) << ')';
-    return out;
-  }
-
-} // namespace geom
-} // namespace carla
-
 static void TransformList(const carla::geom::Transform &self, boost::python::list &list) {
   auto length = boost::python::len(list);
    for (auto i = 0u; i < length; ++i) {
